Do not filterout wifidisplay we are connecting to
During Widi connection we may still received scan results on
p2p interface. If the scan result from the display we are currently
connecting to does not set the bit SessionAvaible in its wfd info
field - which is OK since it has an ongoing connection -, it will
be filterout from the scan result forwarded to upper layer. This
will cause upper layer (MediaRouter) to cancel the on-going connection
assuming the display is no longer available.

This patch does not filterout the display we are currently
connecting from scan result forwarded by WifiDisplayController.

Change-Id: I682a148351cbefdce2b55590c66ac12027c50ba5