[1/4] base: power profiles
Support switching profiles (a set of attributes) based on a
context. A profile can e.g. limit CPU frequency or number
of CPU cores. The device defines what attributes it supports.

KangDroid: This is seems to be in CM, But actually I can't find at all. Anyway Iported to CM, Not sure this will be work or not.
Addition: Move to KangDroid config/symbol

current supported contexts are
1) screen on
2) screen off
3) app specific

The framework part is completely generic and just handles
context switching. Applying the profile attributes is
done in the power HAL

To enable you need

1) add to your device overlay for base
    <bool name="config_powerProfilesSupported">true</bool>

2) implement POWER_HINT_POWER_PROFILE in your devices
power HAL

See https://gerrit.omnirom.org/#/c/11001/ for an example
implementation of the power HAL part

In the future a power profile can contain more generic attributes
like e.g. switching to 2G on screen off

Change-Id: I56ebd5d0507d873859e45a91396e6002357ccefb

Conflicts:
	core/java/android/os/IPowerManager.aidl
	core/java/android/os/PowerManager.java
	core/res/res/values-w320dp/dimens.xml
	core/res/res/values/config.xml
	core/res/res/values/symbols.xml
	services/core/java/com/android/server/am/ActivityStack.java
	services/core/java/com/android/server/power/PowerManagerService.java
	services/core/jni/com_android_server_power_PowerManagerService.cpp

Conflicts:
	core/res/res/values/ackdp_symbols.xml