Themes: Always clone the new theme if non-null
The new theme was only being cloned when shouldUpdateStatusbar()
returned true.  This causes the navigation bar to not be themed
when no components changed that required the status bar to be
recreated.  This patch always clones the new theme so that the
proper navigation bar resources can be loaded.

Change-Id: I8d70c54d3bfcd91141ce40f513744e260fe87639
REF: THEMES-573

Patchset: 4
http://review.cyanogenmod.org/#/c/91705/