systemui: notify everyone that lockscreen is not going to be shown
Instead of just set the mShowing var, we should use the locked method to notify
all the other components that the lockscreen is not going to be shown. Otherwise, the
screen will not be refreshed and will not be drawn.
This happens when PIN set is not enabled (because this one triggers the screen to introduce
the pin). Wihout the PIN set enabled, components are not notified about it.

Change-Id: I949428d47f7ff979afe0e755aa82448ebbd8c598
JIRA: NIGHTLIES-1260
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>

Patchset: 1
http://review.cyanogenmod.org/#/c/99403/